Implements Threads in Decaf instead of using APR, adds autoconf logic to detect pthreads functions.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@814791 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/decaf/util/concurrent/Mutex.h
@@ -53,7 +53,7 @@ namespace concurrent{
         std::list<apr_thread_cond_t*> eventQ;
 
         // Lock Status Members
-        volatile unsigned long lock_owner;
+        volatile long long lock_owner;
         volatile unsigned long lock_count;
 
     private:
